reorganize "billing" module
