+++ liberapay/utils/fake_data.py
@@ -7,7 +7,7 @@ import sys
 from faker import Factory
 from psycopg2 import IntegrityError
 
-from liberapay.billing.exchanges import record_exchange_result, _record_transfer_result
+from liberapay.billing.transactions import record_exchange_result, _record_transfer_result
 from liberapay.constants import D_CENT, DONATION_LIMITS, PERIOD_CONVERSION_RATES
 from liberapay.models.exchange_route import ExchangeRoute
 from liberapay.models.participant import Participant
