+++ www/%username/wallet/payin/card/%back_to.spt
@@ -5,8 +5,9 @@ from decimal import Decimal as D, InvalidOperation
 
 from mangopay.resources import Card, PayIn
 
-from liberapay.billing.exchanges import (
-    charge, record_exchange_result, repr_error, skim_amount, upcharge_card,
+from liberapay.billing.fees import skim_amount, upcharge_card
+from liberapay.billing.transactions import (
+    charge, record_exchange_result, repr_error,
 )
 from liberapay.constants import FEE_PAYIN_CARD, PAYIN_CARD_MIN, KYC_PAYIN_YEARLY_THRESHOLD
 from liberapay.exceptions import InvalidNumber, Redirect
